requirements.txt
requests==2.32.5
python-slugify==8.0.4
markdownify==1.2.2
